requirements.txt
openai
pydub
python-dotenv
streamlit
streamlit-audiorecorder